Add files via upload
Modeling the attractiveness of the implementation of large-scale Agrivoltaics plus battery storage in Germany
The increase of using open-field photovoltaics (PV) can be a large contributor to efforts to mitigate climate change. A constraint for the wide implementation of this technology is the extensive land demand. Large-scale PV faces problems in societal acceptance as well as concerns about the loss of arable land. Increasing population, urbanization, and biofuel production add further pressure to the competition for land.​

​